+++ src/main/java/io/tarantool/driver/api/space/options/UpdateOptions.java
@@ -7,5 +7,5 @@ package io.tarantool.driver.api.space.options;
  * @author Alexey Kuzin
  */
 public interface UpdateOptions<T extends UpdateOptions<T>>
-    extends OperationWithBucketIdOptions<T>, OperationWithTimeoutOptions<T> {
+    extends OperationWithBucketIdOptions<T>, OperationWithTimeoutOptions<T>, OperationWithFieldsOptions<T> {
 }