+++ src/main/java/io/tarantool/driver/core/proxy/CRUDBatchOptions.java
@@ -9,7 +9,7 @@ import java.util.Optional;
  *
  * @author Alexey Kuzin
  */
-public final class CRUDBatchOptions extends CRUDBaseOptions {
+final class CRUDBatchOptions extends CRUDReturnOptions {
 
     public static final String BATCH_STOP_ON_ERROR = "stop_on_error";
     public static final String BATCH_ROLLBACK_ON_ERROR = "rollback_on_error";
@@ -17,30 +17,27 @@ public final class CRUDBatchOptions extends CRUDBaseOptions {
     private <T extends AbstractBuilder<T>>
     CRUDBatchOptions(AbstractBuilder<T> builder) {
         super(builder);
-
         addOption(BATCH_STOP_ON_ERROR, builder.stopOnError);
         addOption(BATCH_ROLLBACK_ON_ERROR, builder.rollbackOnError);
     }
 
-    protected abstract static
-    class AbstractBuilder<T extends AbstractBuilder<T>>
-        extends CRUDBaseOptions.AbstractBuilder<CRUDBatchOptions, T> {
+    protected abstract static class AbstractBuilder<B extends AbstractBuilder<B>>
+            extends CRUDReturnOptions.AbstractBuilder<CRUDBatchOptions, B> {
         private Optional<Boolean> stopOnError = Optional.empty();
         private Optional<Boolean> rollbackOnError = Optional.empty();
 
-        public T withStopOnError(Optional<Boolean> stopOnError) {
+        public B withStopOnError(Optional<Boolean> stopOnError) {
             this.stopOnError = stopOnError;
             return self();
         }
 
-        public T withRollbackOnError(Optional<Boolean> rollbackOnError) {
+        public B withRollbackOnError(Optional<Boolean> rollbackOnError) {
             this.rollbackOnError = rollbackOnError;
             return self();
         }
     }
 
     protected static final class Builder extends AbstractBuilder<Builder> {
-
         @Override
         Builder self() {
             return this;