+++ src/main/java/io/tarantool/driver/core/proxy/CRUDSelectOptions.java
@@ -2,6 +2,7 @@ package io.tarantool.driver.core.proxy;
 
 import io.tarantool.driver.protocol.Packable;
 
+import java.util.List;
 import java.util.Optional;
 
 /**
@@ -18,14 +19,15 @@ final class CRUDSelectOptions extends CRUDBucketIdOptions {
     public static final String SELECT_LIMIT = "first";
     public static final String SELECT_AFTER = "after";
     public static final String SELECT_BATCH_SIZE = "batch_size";
+    public static final String FIELDS = "fields";
 
-    private <T extends AbstractBuilder<T>>
-    CRUDSelectOptions(AbstractBuilder<T> builder) {
+    private <B extends AbstractBuilder<B>> CRUDSelectOptions(AbstractBuilder<B> builder) {
         super(builder);
 
         addOption(SELECT_LIMIT, builder.selectLimit);
         addOption(SELECT_AFTER, builder.after);
         addOption(SELECT_BATCH_SIZE, builder.selectBatchSize);
+        addOption(FIELDS, builder.fields);
     }
 
     /**
@@ -38,6 +40,7 @@ final class CRUDSelectOptions extends CRUDBucketIdOptions {
         private Optional<Long> selectLimit = Optional.empty();
         private Optional<Packable> after = Optional.empty();
         private Optional<Integer> selectBatchSize = Optional.empty();
+        private Optional<List> fields = Optional.empty();
 
         public B withSelectLimit(Optional<Long> selectLimit) {
             this.selectLimit = selectLimit;
@@ -53,13 +56,17 @@ final class CRUDSelectOptions extends CRUDBucketIdOptions {
             this.after = startTuple;
             return self();
         }
+
+        public B withFields(Optional<List> fields) {
+            this.fields = fields;
+            return self();
+        }
     }
 
     /**
      * Concrete Builder implementation for select cluster proxy operation options.
      */
     protected static final class Builder extends AbstractBuilder<Builder> {
-
         @Override
         Builder self() {
             return this;