+++ src/test/java/io/tarantool/driver/integration/proxy/options/ProxySpaceSelectOptionsIT.java
@@ -6,6 +6,7 @@ import io.tarantool.driver.api.TarantoolResult;
 import io.tarantool.driver.api.conditions.Conditions;
 import io.tarantool.driver.api.space.TarantoolSpaceOperations;
 import io.tarantool.driver.api.space.options.proxy.ProxySelectOptions;
+import io.tarantool.driver.api.space.options.SelectOptions;
 import io.tarantool.driver.api.tuple.DefaultTarantoolTupleFactory;
 import io.tarantool.driver.api.tuple.TarantoolTuple;
 import io.tarantool.driver.api.tuple.TarantoolTupleFactory;
@@ -18,12 +19,14 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 /**
  * @author Artyom Dubinin
@@ -120,5 +123,41 @@ public class ProxySpaceSelectOptionsIT extends SharedCartridgeContainer {
         ).get();
         crudSelectOpts = client.eval("return crud_select_opts").get();
         assertEquals(customRequestTimeout, ((HashMap) crudSelectOpts.get(0)).get("timeout"));
+    }
+
+    @Test
+    public void withFieldsTest() throws ExecutionException, InterruptedException {
+        TarantoolSpaceOperations<TarantoolTuple, TarantoolResult<TarantoolTuple>> profileSpace =
+                client.space(TEST_SPACE_NAME);
+
+        TarantoolTuple tarantoolTuple;
+
+        tarantoolTuple = tupleFactory.create(1, null, "FIO", 50, 100);
+        profileSpace.insert(tarantoolTuple).get();
+
+        Conditions conditions = Conditions.equals(PK_FIELD_NAME, 1);
+
+        // without fields
+        TarantoolResult<TarantoolTuple> selectResult = profileSpace.select(conditions).get();
+        assertEquals(1, selectResult.size());
+
+        TarantoolTuple tuple = selectResult.get(0);
+        assertEquals(5, tuple.size());
+        assertEquals(1, tuple.getInteger(0));
+        assertNotNull(tuple.getInteger(1)); //bucket_id
+        assertEquals("FIO", tuple.getString(2));
+        assertEquals(50, tuple.getInteger(3));
+        assertEquals(100, tuple.getInteger(4));
+
+        // with fields
+        SelectOptions options = ProxySelectOptions.create().withFields(Arrays.asList("profile_id", "age"));
+        selectResult = profileSpace.select(conditions, options).get();
+        assertEquals(1, selectResult.size());
+
+        tuple = selectResult.get(0);
+        assertEquals(2, tuple.size());
+        assertEquals(1, tuple.getInteger(0));
+        assertEquals(1, tuple.getInteger("profile_id"));
+        assertEquals(50, tuple.getInteger("age"));
     }
 }