+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Added `fields` option to ProxySpace ([#236](https://github.com/tarantool/cartridge-java/pull/236))
 - Move metadata parsing to separate converters ([#325](https://github.com/tarantool/cartridge-java/pull/325))
 - Parse metadata from crud response ([#272](https://github.com/tarantool/cartridge-java/pull/272))
 - Use netty part dependencies instead of netty-all ([#295](https://github.com/tarantool/cartridge-java/issues/295))